upodate

+++ 0x02-functions_nested_loops/104-fibonacci.c
@@ -6,19 +6,19 @@
 int main(void)
 {
 	int i;
-	unsigned long int f[98];
+	unsigned long long int f[98];
 
 	f[0] = 1;
 	f[1] = 2;
-	printf("%lu, %lu, ", f[0], f[1]);
+	printf("%llu, %llu, ", f[0], f[1]);
 
 	for (i = 2; i < 98; i++)
 	{
 		f[i] = f[i - 1] + f[i - 2];
 		if (i == 97)
-			printf("%lu\n", f[i]);
+			printf("%llu\n", f[i]);
 		else
-			printf("%lu, ", f[i]);
+			printf("%llu, ", f[i]);
 	}
 	return (0);
 }
